Added water, parsing of water, printStatistics method, and much more

+++ src/game/AirWarrior.java
@@ -1,7 +1,7 @@
 package game;
 public class AirWarrior extends Warrior implements WarriorTypeInterface {
 
-    public AirWarrior(Position position, int id, int age, double health, int offense, int defense, int invSize, String moves) {
+    public AirWarrior(Position position, int id, int age, double health, double offense, double defense, int invSize, String moves) {
         super(position, id, age, health, offense, defense, "Air", invSize, moves);
     }
