+++ src/game/WaterWarrior.java
@@ -1,7 +1,7 @@
 package game;
 public class WaterWarrior extends Warrior implements WarriorTypeInterface{
 
-    public WaterWarrior(Position position, int id, int age, double health, int offense, int defense, int invSize, String moves) {
+    public WaterWarrior(Position position, int id, int age, double health, double offense, double defense, int invSize, String moves) {
         super(position, id, age, health, offense, defense, "Water", invSize, moves);
     }
